主要是写了一下staticImage和HeightMap，顺便把图片都添加到了项目中

+++ shoot_test/shoot_test/Model.h
@@ -48,7 +48,7 @@ public :
     {
         float m_vertexNormals[3][3];
         float m_s[3], m_t[3];
-        float m_vertexIndices[3];
+        int m_vertexIndices[3];
     };
 
     struct Vertex
